+++ merchantops-api/src/test/java/com/renda/merchantops/api/service/TicketCommandServiceTest.java
@@ -47,6 +47,9 @@ class TicketCommandServiceTest {
     @Mock
     private UserRepository userRepository;
 
+    @Mock
+    private AuditEventService auditEventService;
+
     @InjectMocks
     private TicketCommandService ticketCommandService;